requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
langchain==0.2.5
langchain-google-genai==1.0.6
requests==2.32.3
streamlit==1.36.0
langgraph==0.0.69
thefuzz[speedup]==0.22.1
langchain-core==0.2.7
dateparser==1.2.0
